Add GitHub Actions CI, Vitest, and blog markdown extraction
